# api/swagger_server/models/user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class User(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: int=None, name: str=None, email: str=None, password: str=None, logged: bool=False, type: str=None):  # noqa: E501
        """User - a model defined in Swagger

        :param id: The id of this User.  # noqa: E501
        :type id: int
        :param name: The name of this User.  # noqa: E501
        :type name: str
        :param email: The email of this User.  # noqa: E501
        :type email: str
        :param password: The password of this User.  # noqa: E501
        :type password: str
        :param logged: The logged of this User.  # noqa: E501
        :type logged: bool
        :param type: The type of this User.  # noqa: E501
        :type type: str
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'email': str,
            'password': str,
            'logged': bool,
            'type': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'email': 'email',
            'password': 'password',
            'logged': 'logged',
            'type': 'type'
        }

        self._id = id
        self._name = name
        self._email = email
        self._password = password
        self._logged = logged
        self._type = type

    @classmethod
    def from_dict(cls, dikt) -> 'User':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The User of this User.  # noqa: E501
        :rtype: User
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this User.


        :return: The id of this User.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this User.


        :param id: The id of this User.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this User.


        :return: The name of this User.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this User.


        :param name: The name of this User.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def email(self) -> str:
        """Gets the email of this User.


        :return: The email of this User.
        :rtype: str
        """
        return self._email

    @email.setter
    def email(self, email: str):
        """Sets the email of this User.


        :param email: The email of this User.
        :type email: str
        """

        self._email = email

    @property
    def password(self) -> str:
        """Gets the password of this User.


        :return: The password of this User.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this User.


        :param password: The password of this User.
        :type password: str
        """

        self._password = password

    @property
    def logged(self) -> bool:
        """Gets the logged of this User.


        :return: The logged of this User.
        :rtype: bool
        """
        return self._logged

    @logged.setter
    def logged(self, logged: bool):
        """Sets the logged of this User.


        :param logged: The logged of this User.
        :type logged: bool
        """

        self._logged = logged

    @property
    def type(self) -> str:
        """Gets the type of this User.


        :return: The type of this User.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this User.


        :param type: The type of this User.
        :type type: str
        """
        allowed_values = ["admin", "user"]  # noqa: E501
        if type not in allowed_values:
            raise ValueError(
                "Invalid value for `type` ({0}), must be one of {1}"
                .format(type, allowed_values)
            )

        self._type = type