# api/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST API definition for EVENTually project, a simple user friendly\
    \ event manager to host/manage all the events in hte the organization."
  version: "1.0.0"
  title: "EVENTually"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "sgorai@rbbn.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
schemes:
- "https"
- "http"
paths:
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "create_user"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Info"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/Info"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      operationId: "login_user"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Username and password"
        required: true
        schema:
          $ref: "#/definitions/Login_details"
      responses:
        "200":
          description: "Successfully logged in"
        "400":
          description: "Bad Request"
        "500":
          description: "Server Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/logout/{userid}:
    post:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      operationId: "logout_user"
      produces:
      - "application/json"
      parameters:
      - name: "userid"
        in: "path"
        description: "Username that needs to be logged out."
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successfully user logged out"
        "400":
          description: "Bad Request"
        "500":
          description: "Server Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/{userid}:
    get:
      tags:
      - "user"
      summary: "Get user by user id"
      description: ""
      operationId: "get_user_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "userid"
        in: "path"
        description: "The id that needs to be fetched."
        required: true
        type: "integer"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Info"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/Info"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
    put:
      tags:
      - "user"
      summary: "Update user"
      description: "This can only be done by the logged in user."
      operationId: "modify_user_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "userid"
        in: "path"
        description: "id that need to be updated"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Update user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Info"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/Info"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "delete_user_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "userid"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Info"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/Info"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
definitions:
...
  Login_details:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      password:
        type: "string"
    example:
      username: "username"
      password: "password"
...
  User:
    type: "object"
    required:
    - "name"
    - "type"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      logged:
        type: "boolean"
        default: false
      type:
        type: "string"
        enum:
        - "admin"
        - "user"
    example:
      password: "password"
      logged: false
      name: "name"
      id: 0
      type: "admin"
      email: "email"
  Info:
    type: "object"
    properties:
      critical:
        $ref: "#/definitions/Info_critical"
      error:
        $ref: "#/definitions/Info_error"
      warning:
        $ref: "#/definitions/Info_warning"
      info:
        $ref: "#/definitions/Info_info"
    description: "Contains information about the request which could include warnings,\
      \ error messages, and information about any side effects of some operations\
      \ such as which dashboards were added or removed.\n"
    example:
      critical:
        details:
        - "details"
        - "details"
      warning:
        details:
        - "details"
        - "details"
      error:
        details:
        - "details"
        - "details"
      info:
        details:
        - "details"
        - "details"
  Info_critical:
    properties:
      details:
        type: "array"
        items:
          type: "string"
    description: "Contains information about a critical error that has occurred that\
      \ could affect the availability of the service.\n"
    example:
      details:
      - "details"
      - "details"
  Info_error:
    properties:
      details:
        type: "array"
        items:
          type: "string"
    description: "Contains information about a error that has occurred during the\
      \ execution of the requested operation.\n"
    example:
      details:
      - "details"
      - "details"
  Info_warning:
    properties:
      details:
        type: "array"
        items:
          type: "string"
    description: "Contains information that does not affect the results of the requested\
      \ operation but which is out of the ordinary and may require user action or\
      \ investigation. For example, a warning could be used to identify the use of\
      \ deprecated REST API calls.\n"
    example:
      details:
      - "details"
      - "details"
  Info_info:
    properties:
      details:
        type: "array"
        items:
          type: "string"
    description: "Contains details of the requested operation that would be of interest\
      \ to the user. For example, it could include a list of items that have been\
      \ deleted when a containing object is deleted.\n"
    example:
      details:
      - "details"
      - "details"